refactor: add base transformer
