+++ src/customizer/lang_processor/aspect_transformer.py
@@ -1,17 +1,12 @@
-from lark import Token, Transformer
-
 from customizer.aspect.advice import Advice
 from customizer.aspect.aspect import Aspect
 from customizer.aspect.basic_aspect import BasicAspect
 from customizer.method.method import Method
 from customizer.pointcut.execution import Execution
-from customizer.pointcut.func_signature import FuncSignature
+from customizer.lang_processor.base_transformer import BaseTransformer
 
 
-class AspectTransformer(Transformer):
-    def start(self, tree):
-        return tree
-
+class AspectTransformer(BaseTransformer):
     ######################## aspect ########################
     def basic_aspect(self, tree):
         """aspect Baz {
@@ -57,68 +52,3 @@ class AspectTransformer(Transformer):
 
     def infunc(self, tree):
         return {"name": "infunc", "arg": tree[0]}
-
-    ######################## method ########################
-    def method(self, tree):
-        """
-        e.g.) tree[FuncSignature, str] → int foo(char) { ... }
-        """
-        signature, body = tree[0], tree[1]
-        return Method(signature, body)
-
-    ######################## func ########################
-
-    def func_signature(self, tree):
-        """
-        e.g.) tree[int, foo, char] -> int foo(char)
-        """
-        type, name, args = str(tree[0]), tree[1], tree[2]
-        return FuncSignature(type, name, args)
-
-    def func_name(self, tree):
-        """
-        e.g.) set_foo
-        """
-        return str(tree[0])
-
-    def func_args(self, tree):
-        """
-        return:
-            e.g.) ['int', 'float', 'char']
-        """
-        return [str(t) for t in tree]
-
-    def body(self, tree):
-        """
-        e.g.)
-        foo();
-        if() {
-            bar();
-        }
-        → [foo();, if() {, bar();, }]
-        """
-        lines: list[str] = []
-        for t in tree:
-            if isinstance(t, Token):
-                lines.append(t.value)
-            else:
-                lines += t
-        return lines
-
-    def block(self, tree):
-        """
-        e.g.)
-        if() {
-            if() {
-                foo();
-            }
-         }
-        →[if(), {, if(), foo();, }, }]
-        """
-        lines: list[str] = []
-        for t in tree:
-            if isinstance(t, Token):
-                lines.append(t.value)
-            else:
-                lines += t
-        return lines
