+++ src/customizer/lang_processor/inheritance_transformer.py
@@ -1,5 +1,3 @@
-from lark import Token, Transformer
-
 from customizer.primitive_aspect.stringified_aspect import StringifiedAspect
 from customizer.primitive_aspect.abstract_aspect import AbstractAspect
 from customizer.primitive_aspect.primitive_aspect import PrimitiveAspect
@@ -7,12 +5,10 @@ from customizer.primitive_aspect.concrete_aspect import ConcreteAspect
 from customizer.primitive_aspect.constructor import Constructor
 from customizer.method.method import Method
 from customizer.pointcut.func_signature import FuncSignature
+from customizer.lang_processor.base_transformer import BaseTransformer
 
 
-class InheritanceTransformer(Transformer):
-    def start(self, tree):
-        return tree
-
+class InheritanceTransformer(BaseTransformer):
     ######################## aspect ########################
     def aspect_container(self, tree):
         return tree[0]
@@ -102,62 +98,3 @@ class InheritanceTransformer(Transformer):
     ######################## method ########################
     def abstract_method(self, tree) -> FuncSignature:
         return tree[1]
-
-    def method(self, tree):
-        signature, body = tree[0], tree[1]
-        return Method(signature, body)
-
-    def func_signature(self, tree):
-        """
-        e.g.) tree[int, foo, char] -> int foo(char)
-        """
-        type, name, args = str(tree[0]), tree[1], tree[2]
-        return FuncSignature(type, name, args)
-
-    def func_name(self, tree):
-        """
-        e.g.) set_foo
-        """
-        return str(tree[0])
-
-    def func_args(self, tree):
-        """
-        return:
-            e.g.) ['int', 'float', 'char']
-        """
-        return [str(t) for t in tree]
-
-    def body(self, tree):
-        """
-        e.g.)
-        foo();
-        if() {
-            bar();
-        }
-        → [foo();, if() {, bar();, }]
-        """
-        lines: list[str] = []
-        for t in tree:
-            if isinstance(t, Token):
-                lines.append(t.value)
-            else:
-                lines += t
-        return lines
-
-    def block(self, tree):
-        """
-        e.g.)
-        if() {
-            if() {
-                foo();
-            }
-         }
-        →[if(), {, if(), foo();, }, }]
-        """
-        lines: list[str] = []
-        for t in tree:
-            if isinstance(t, Token):
-                lines.append(t.value)
-            else:
-                lines += t
-        return lines
